refactor: :recycle: derive error handling in AuthForm and OutsideProviders to AuthErrorModal
