+++ components/Auth/AuthForm/AuthForm.tsx
@@ -1,9 +1,8 @@
-import { useEffect } from 'react'
 import { useForm, Controller } from 'react-hook-form'
 import * as yup from 'yup'
 import YupPassword from 'yup-password'
 import { yupResolver } from '@hookform/resolvers/yup'
-import { Input, Typography, Modal } from 'antd'
+import { Input, Typography } from 'antd'
 import {
   SForm,
   Btn,
@@ -13,7 +12,6 @@ import {
   ForgotPassword,
 } from './AuthForm.styles'
 import OutsideProvidersAuth from '../OutsideProvidersAuth/OutsideProvidersAuth'
-import { getMessageFromErrorCode } from '@/utils/getMessageFromAuthError'
 import capitalizeWord from '@/utils/capitalizeWord'
 
 YupPassword(yup)
@@ -36,57 +34,18 @@ type AuthFormProps = {
   setAuthMethod: (method: 'login' | 'signup' | 'password-reset') => void
   onSubmit: (data: AuthFormInputsData) => void
   schema: yup.AnyObjectSchema
-  errorCode: string
-  setErrorCode: (arg0: string) => void
 }
 
-const AuthForm = ({
-  title,
-  inputs,
-  btnText,
-  setAuthMethod,
-  onSubmit,
-  schema,
-  errorCode,
-  setErrorCode,
-}: AuthFormProps) => {
+const AuthForm = ({ title, inputs, btnText, setAuthMethod, onSubmit, schema }: AuthFormProps) => {
   const {
     control,
     handleSubmit,
     formState: { errors },
   } = useForm<AuthFormInputsData>({ resolver: yupResolver(schema) })
-  const [modal, contextHolder] = Modal.useModal()
-
-  const onFormSubmit = (data: AuthFormInputsData) => {
-    if (errorCode) {
-      showModal()
-    } else {
-      onSubmit(data)
-    }
-  }
-
-  const showModal = () => {
-    const errorMessage = getMessageFromErrorCode(errorCode)
-
-    modal.error({
-      title: 'Error!',
-      content: <ErrorMessage>{errorMessage}</ErrorMessage>,
-      centered: true,
-      onOk: () => {
-        setErrorCode('')
-      },
-    })
-  }
-
-  useEffect(() => {
-    if (errorCode) {
-      showModal()
-    }
-  })
 
   return (
     <>
-      <SForm onSubmit={handleSubmit(onFormSubmit)}>
+      <SForm onSubmit={handleSubmit(onSubmit)}>
         <h2>{title}</h2>
         {inputs.map((input) => {
           const { name, type } = input
@@ -142,7 +101,6 @@ const AuthForm = ({
           )}
         </HighlightedSpanContainer>
       </SForm>
-      {contextHolder}
     </>
   )
 }
